Solved Coupon bug and addto cart count bug
Once a coupon is applied that user will be restricted from using the same coupon again but if we refresh the page , the coupon offer will be reversed but still the user won't be able to apply the same coupon . I fixed this problem by restricting the user only if the order is successfully placed.
Second I provided the user details in the razorpay payment.